Trying to make a webpage to write to Firebase
addRequest takes three arguments, name, description, and an array of Tags, and creates a request Document on Firebase.  In the index.html file, there is a basic interface with three text boxes to enter the three aforementioned parameters, with a "submit" button.  Clicking the submit button calls onClick(), which takes the inputs from the text fields and formats them for addRequest().  However, there is some error with calling the addRequest() method, since it is located in a different file.  addRequest() must be in the index.js file since that's where all the other firebase stuff are.  The part to "import" the js file to the .html file is located above the head, in the <script>.  The code should run with npm start, but of the button is clicked with any input in the text fields, then an error message saying addRequest is not defined, or something similar, is printed in console.log.