Fixed the error; now writes to database.
Four textboxes allow the submission of a request to the Firebase.  The first blank is a title, second is description, third is tags separated by commas, and fourth is bounty.  There is no confirmation message when a request is submitted (yet).  Also, from the Firebase view, all the request document names are hashed based on time; this allows for two requests with the same title.